requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
Authlib==1.2.1
bcrypt==4.0.1
boto3==1.28.3
botocore==1.31.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
cryptography==41.0.2
exceptiongroup==1.1.3
fastapi==0.100.0
GeoAlchemy2==0.14.1
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
packaging==23.2
passlib==1.7.4
protobuf==3.20.3
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.0.2
pydantic_core==2.1.2
PyJWT==2.7.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
